requirements.txt
arrow==1.2.2
asgiref==3.5.1
attrs==21.4.0
backports.zoneinfo==0.2.1
binaryornot==0.4.4
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cookiecutter==1.7.3
coverage==6.3.2
crispy-tailwind==0.5.0
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.4
django-browser-reload==1.3.0
django-crispy-forms==1.14.0
django-phonenumber-field==6.1.0
django-tailwind==3.1.1
djangorestframework==3.13.1
djlint==1.0.1
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
mypy==0.950
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.48
platformdirs==2.5.2
pluggy==1.0.0
poyo==0.5.0
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.8
pytest==7.1.2
pytest-cov==3.0.0
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
text-unidecode==1.3
tomli==2.0.1
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
